+++ app/src/main/java/com/example/weather/repo/Repository.kt
@@ -6,10 +6,10 @@ import com.example.weather.model.WeatherData
 object Repository {
     private val apiservice=Network.getApiServices()
 
-    suspend fun getWeatherDetails(lat:Float,lon:Float):WeatherData {
+    suspend fun getWeatherDetails(lat:Double,lon:Double):WeatherData {
         return apiservice.getWeatherDetail(lat,lon,Network.apikey)
     }
-    suspend fun getAllWeatherDetail(lat: Float,lon: Float):AllWeatherData{
+    suspend fun getAllWeatherDetail(lat: Double,lon: Double):AllWeatherData{
         return apiservice.getAllDayWeatherDetail(lat,lon,"day",Network.apikey)
     }
 }
